reduce shader variable precision, fix some undefined behavior

+++ go.graphics/src/main/resources/color.frag
@@ -1,6 +1,6 @@
 #version 100
 
-uniform highp vec4 color;
+uniform mediump vec4 color;
 
 void main() {
 	gl_FragColor = color;
